Migrate S2432 to eslint implementation
