+++ eslint-bridge/src/rules/main.ts
@@ -122,6 +122,7 @@ import { rule as noRedundantParentheses } from './no-redundant-parentheses';
 import { rule as noReferenceError } from './no-reference-error';
 import { rule as noReferrerPolicy } from './no-referrer-policy';
 import { rule as noRequireOrDefine } from './no-require-or-define';
+import { rule as noReturnInSetter } from './no-return-in-setter';
 import { rule as noReturnTypeAny } from './no-return-type-any';
 import { rule as noTab } from './no-tab';
 import { rule as noTryPromise } from './no-try-promise';
@@ -282,6 +283,7 @@ ruleModules['no-redundant-parentheses'] = noRedundantParentheses;
 ruleModules['no-reference-error'] = noReferenceError;
 ruleModules['no-referrer-policy'] = noReferrerPolicy;
 ruleModules['no-require-or-define'] = noRequireOrDefine;
+ruleModules['no-return-in-setter'] = noReturnInSetter;
 ruleModules['no-return-type-any'] = noReturnTypeAny;
 ruleModules['no-tab'] = noTab;
 ruleModules['no-try-promise'] = noTryPromise;
